fix: fix blurhash not being rendered for unsplash images
